+++ GrowableGiantCrops/HarmonyPatches/Niceties/PatchesForSObject.cs
@@ -2,6 +2,7 @@
 
 using GrowableGiantCrops.Framework;
 using GrowableGiantCrops.Framework.InventoryModels;
+using GrowableGiantCrops.HarmonyPatches.Compat;
 
 using HarmonyLib;
 
@@ -30,22 +31,24 @@ internal static class PatchesForSObject
     [SuppressMessage("StyleCop.CSharp.NamingRules", "SA1313:Parameter names should begin with lower-case letter", Justification = "Harmony convention.")]
     private static void PostfixPlacement(SObject __instance, GameLocation location, int x, int y)
     {
-        if (__instance?.bigCraftable?.Value == true)
+        if (__instance?.bigCraftable?.Value == true
+            && location.Objects.TryGetValue(new Vector2(x / Game1.tileSize, y / Game1.tileSize), out SObject? placed))
         {
             if (__instance.Name == "Slime Ball")
             {
-                __instance.modData?.SetBool(ModDataKey, true);
-                __instance.TileLocation = new Vector2(x / Game1.tileSize, y / Game1.tileSize);
+                placed.modData.Remove(SlimeProduceCompat.SlimeBall);
+                placed.modData.SetBool(ModDataKey, true);
+                placed.TileLocation = new Vector2(x / Game1.tileSize, y / Game1.tileSize);
             }
             else if (__instance.Name == "Mushroom Box")
             {
-                __instance.modData?.SetBool(ModDataMiscObject, true);
-                __instance.TileLocation = new Vector2(x / Game1.tileSize, y / Game1.tileSize);
+                placed.modData?.SetBool(ModDataMiscObject, true);
+                placed.TileLocation = new Vector2(x / Game1.tileSize, y / Game1.tileSize);
             }
             else if (__instance.ParentSheetIndex == 78)
             {
-                __instance.modData?.SetBool(ModDataMiscObject, true);
-                __instance.Fragility = SObject.fragility_Removable;
+                placed.modData?.SetBool(ModDataMiscObject, true);
+                placed.Fragility = SObject.fragility_Removable;
             }
         }
         else if (__instance?.bigCraftable?.Value == false)