+++ GrowableGiantCrops/docs/changelog.md
@@ -2,11 +2,13 @@
 ===========
 
 TODO:
-   - Enchantments?
-   - Check map modifications. Which ones do I want to yeet?
    - Custom resource clumps and custom ore nodes?
    - Fix up grange draw?
    - Maybe put grass in the furniture catalogue?
+
+### Version 0.1.1
+* Fixed compat with Slime Produce
+* Fixed small items not getting their moddata marked correctly.
 
 ### Version 0.1.0