+++ HighlightEmptyMachines/HighlightEmptyMachines/docs/Changelog.md
@@ -3,6 +3,9 @@
 
 TODO: optionally highlight **buildings** that have empty machines within them?
 
+### Version 0.2.3
+* fix PFM integration.
+
 ### Version 0.2.2
 * Recompiled for new AtraCore, my apologies!
 * By popular request, allow you to disable the pulsing.